fixed LineaSearch name and impleneted tests
